requirements.txt
# MBA论文AIGC检测系统依赖

# 核心依赖
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 文档处理
pypdf>=3.0.0
python-docx>=0.8.11

# API请求（可选，用于SiliconFlow/ollama）
requests>=2.28.0

# 工具
pytest>=7.0.0
